Command prefix `[n]` tells the interleaver how many CPU cycles the command is executed
For example, "[4] mov ax,[bx]" means that the delay between this and the next command in the same stream is 4 cycles (default value is 1). This allows the interleaver to put after a high-latency command more commands from other streams.

This also allows more precise, semi-manual control of the interleaving. In particular, one can use the `[0]` prefix to ensure that commands will be issued back-to-back (and thus immediately free all registers occupied by variables used only inside these commands).

You can find a creative use of this feature in  examples/lz_match_finder.neasm